deleted duplicate code
