+++ frontend/src/Game_Related_Components/DeckBuildingComponents/DeckBuilderPage.css
@@ -28,4 +28,7 @@
 .Decks {
   display: flex;
 }
+.PlayersDecks{
+  margin-left:5%;
+}
